+++ scripts/build-tumbleweed-kernel.sh
@@ -85,7 +85,7 @@ on_exit() {
 	local result=${1}
 
 	if [ -d ${tmp_dir} ]; then
-		rm -rf ${tmp_dir}
+		rm -rf "${tmp_dir:?}"
 	fi
 
 	local end_time=${SECONDS}
