+++ scripts/prepare-ld-scripts.sh
@@ -69,7 +69,7 @@ on_exit() {
 	local result=${1}
 
 	if [[ -d ${tmp_dir} ]]; then
-		rm -rf ${tmp_dir}
+		rm -rf "${tmp_dir:?}"
 	fi
 
 	set +x
