+++ scripts/tftp-upload.sh
@@ -129,7 +129,7 @@ on_exit() {
 	local result=${1}
 
 	if [[ ${tmp_dir} && -d ${tmp_dir} ]]; then
-		rm -rf ${tmp_dir}
+		rm -rf "${tmp_dir:?}"
 	fi
 
 	echo "${script_name}: ${result}" >&2
